add tests for all interfaces
