+++ samples/Apache.IoTDB.Samples/SessionPoolTest.AlignedRecord.cs
@@ -3,13 +3,11 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Apache.IoTDB.DataStructure;
-
 namespace Apache.IoTDB.Samples
 {
     public partial class SessionPoolTest
     {
-
-        public async Task TestInsertRecord()
+        public async Task TestInsertAlignedRecord()
         {
             var session_pool = new SessionPool(host, port, pool_size);
             int status;
@@ -19,102 +17,34 @@ namespace Apache.IoTDB.Samples
             System.Diagnostics.Debug.Assert(session_pool.IsOpen());
             status = await session_pool.DeleteStorageGroupAsync(test_group_name);
 
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[1]), TSDataType.TEXT,
-                TSEncoding.PLAIN, Compressor.UNCOMPRESSED);
+            string prefixPath = string.Format("{0}.{1}", test_group_name, test_device);
+            var measurements = new List<string> { test_measurements[1], test_measurements[2], test_measurements[3] };
+            var types = new List<TSDataType> { TSDataType.TEXT, TSDataType.BOOLEAN, TSDataType.INT32 };
+            var encodings = new List<TSEncoding> { TSEncoding.PLAIN, TSEncoding.PLAIN, TSEncoding.PLAIN };
+            var compressors = new List<Compressor> { Compressor.UNCOMPRESSED, Compressor.UNCOMPRESSED, Compressor.UNCOMPRESSED };
+            status = await session_pool.CreateAlignedTimeseriesAsync(prefixPath, measurements, types, encodings, compressors);
+            System.Diagnostics.Debug.Assert(status == 0);
 
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[2]),
-                TSDataType.BOOLEAN, TSEncoding.PLAIN, Compressor.UNCOMPRESSED);
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[3]),
-                TSDataType.INT32, TSEncoding.PLAIN, Compressor.UNCOMPRESSED);
-            System.Diagnostics.Debug.Assert(status == 0);
             var measures = new List<string>
-                {test_measurements[1], test_measurements[2], test_measurements[3]};
+                { test_measurements[1], test_measurements[2], test_measurements[3] };
             var values = new List<object> { "test_text", true, (int)123 };
             var tasks = new List<Task<int>>();
             var start_ms = DateTime.Now.Ticks / 10000;
             for (var timestamp = 1; timestamp <= fetch_size * processed_size; timestamp++)
             {
                 var rowRecord = new RowRecord(timestamp, values, measures);
-                var task = session_pool.InsertRecordAsync(
+                var task = session_pool.InsertAlignedRecordAsync(
                     string.Format("{0}.{1}", test_group_name, test_device), rowRecord);
                 tasks.Add(task);
             }
-
             Task.WaitAll(tasks.ToArray());
             var end_ms = DateTime.Now.Ticks / 10000;
-            Console.WriteLine(string.Format("total insert aligned record time is {0}", end_ms - start_ms));
+            Console.WriteLine(string.Format("total insert record time is {0}", end_ms - start_ms));
             status = await session_pool.DeleteStorageGroupAsync(test_group_name);
             await session_pool.Close();
-            Console.WriteLine("TestInsertRecordAsync Passed");
+            Console.WriteLine("TestInsertAlignedRecordAsync Passed");
         }
-        public async Task TestInsertStrRecord()
-        {
-            var session_pool = new SessionPool(host, port, pool_size);
-            var status = 0;
-            await session_pool.Open(false);
-            if (debug) session_pool.OpenDebugMode();
-
-            System.Diagnostics.Debug.Assert(session_pool.IsOpen());
-            await session_pool.DeleteStorageGroupAsync(test_group_name);
-
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[1]),
-                TSDataType.INT32, TSEncoding.PLAIN, Compressor.UNCOMPRESSED);
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[2]),
-                TSDataType.INT32, TSEncoding.PLAIN, Compressor.UNCOMPRESSED);
-            System.Diagnostics.Debug.Assert(status == 0);
-
-            var measures = new List<string> { test_measurements[1], test_measurements[2] };
-            var values = new List<object> { (int)1, (int)2 };
-            var rowRecord = new RowRecord(1, values, measures);
-            status = await session_pool.InsertRecordAsync(
-                string.Format("{0}.{1}", test_group_name, test_device), rowRecord);
-            System.Diagnostics.Debug.Assert(status == 0);
-            var res = await session_pool.ExecuteQueryStatementAsync(
-                "select * from " + string.Format("{0}.{1}", test_group_name, test_device) + " where time<2");
-            res.ShowTableNames();
-            while (res.HasNext()) Console.WriteLine(res.Next());
-
-            await res.Close();
-
-            var tasks = new List<Task<int>>();
-            // large data test
-            var rowRecords = new List<RowRecord>() { };
-            for (var timestamp = 2; timestamp <= fetch_size * processed_size; timestamp++)
-                rowRecords.Add(new RowRecord(timestamp, values, measures));
-
-            for (var timestamp = 2; timestamp <= fetch_size * processed_size; timestamp++)
-            {
-                var task = session_pool.InsertRecordAsync(
-                    string.Format("{0}.{1}", test_group_name, test_device), rowRecords[timestamp - 2]);
-                tasks.Add(task);
-            }
-
-            Task.WaitAll(tasks.ToArray());
-            res = await session_pool.ExecuteQueryStatementAsync(
-                "select * from " + string.Format("{0}.{1}", test_group_name, test_device));
-            var res_count = 0;
-            while (res.HasNext())
-            {
-                res.Next();
-                res_count += 1;
-            }
-
-            await res.Close();
-            Console.WriteLine(res_count + " " + fetch_size * processed_size);
-            System.Diagnostics.Debug.Assert(res_count == fetch_size * processed_size);
-            await session_pool.DeleteStorageGroupAsync(test_group_name);
-            await session_pool.Close();
-            Console.WriteLine("TestInsertStrRecord Passed!");
-        }
-        public async Task TestInsertRecords()
+        public async Task TestInsertAlignedRecords()
         {
             var session_pool = new SessionPool(host, port, pool_size);
             await session_pool.Open(false);
@@ -123,34 +53,38 @@ namespace Apache.IoTDB.Samples
             System.Diagnostics.Debug.Assert(session_pool.IsOpen());
             var status = 0;
             await session_pool.DeleteStorageGroupAsync(test_group_name);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[1]),
-                TSDataType.BOOLEAN, TSEncoding.PLAIN, Compressor.SNAPPY);
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[2]),
-                TSDataType.INT32, TSEncoding.PLAIN, Compressor.SNAPPY);
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[3]),
-                TSDataType.INT64, TSEncoding.PLAIN, Compressor.SNAPPY);
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[4]),
-                TSDataType.DOUBLE, TSEncoding.PLAIN, Compressor.SNAPPY);
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[5]),
-                TSDataType.FLOAT, TSEncoding.PLAIN, Compressor.SNAPPY);
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[6]), TSDataType.TEXT,
-                TSEncoding.PLAIN, Compressor.SNAPPY);
+
+            string prefixPath = string.Format("{0}.{1}", test_group_name, test_device);
+            var measurement_lst = new List<string>()
+            {
+                test_measurements[1],
+                test_measurements[2],
+                test_measurements[3],
+                test_measurements[4],
+                test_measurements[5],
+                test_measurements[6]
+            };
+            var data_type_lst = new List<TSDataType>()
+            {
+                TSDataType.BOOLEAN, TSDataType.INT32, TSDataType.INT64, TSDataType.DOUBLE, TSDataType.FLOAT,
+                TSDataType.TEXT
+            };
+            var encoding_lst = new List<TSEncoding>()
+            {
+                TSEncoding.PLAIN, TSEncoding.PLAIN, TSEncoding.PLAIN, TSEncoding.PLAIN, TSEncoding.PLAIN,
+                TSEncoding.PLAIN
+            };
+            var compressor_lst = new List<Compressor>()
+            {
+                Compressor.SNAPPY, Compressor.SNAPPY, Compressor.SNAPPY, Compressor.SNAPPY, Compressor.SNAPPY,
+                Compressor.SNAPPY
+            };
+            status = await session_pool.CreateAlignedTimeseriesAsync(prefixPath, measurement_lst, data_type_lst, encoding_lst,
+                compressor_lst);
             System.Diagnostics.Debug.Assert(status == 0);
 
             var device_id = new List<string>() { };
             for (var i = 0; i < 3; i++) device_id.Add(string.Format("{0}.{1}", test_group_name, test_device));
-
             var measurements_lst = new List<List<string>>() { };
             measurements_lst.Add(new List<string>() { test_measurements[1], test_measurements[2] });
             measurements_lst.Add(new List<string>()
@@ -182,7 +116,7 @@ namespace Apache.IoTDB.Samples
                 rowRecords.Add(rowRecord);
             }
 
-            status = await session_pool.InsertRecordsAsync(device_id, rowRecords);
+            status = await session_pool.InsertAlignedRecordsAsync(device_id, rowRecords);
             System.Diagnostics.Debug.Assert(status == 0);
             var res = await session_pool.ExecuteQueryStatementAsync(
                 "select * from " + string.Format("{0}.{1}", test_group_name, test_device) + " where time<10");
@@ -203,7 +137,7 @@ namespace Apache.IoTDB.Samples
                     new List<string>() { test_measurements[1], test_measurements[2] }));
                 if (timestamp % fetch_size == 0)
                 {
-                    tasks.Add(session_pool.InsertRecordsAsync(device_id, rowRecords));
+                    tasks.Add(session_pool.InsertAlignedRecordsAsync(device_id, rowRecords));
                     device_id = new List<string>() { };
                     rowRecords = new List<RowRecord>() { };
                 }
@@ -225,12 +159,13 @@ namespace Apache.IoTDB.Samples
             Console.WriteLine(res_count + " " + fetch_size * processed_size);
             System.Diagnostics.Debug.Assert(res_count == record_count);
             System.Diagnostics.Debug.Assert(status == 0);
+
             status = await session_pool.DeleteStorageGroupAsync(test_group_name);
             System.Diagnostics.Debug.Assert(status == 0);
             await session_pool.Close();
-            Console.WriteLine("TestInsertRecords Passed!");
+            Console.WriteLine("TestInsertAlignedRecords Passed!");
         }
-        public async Task TestInsertRecordsOfOneDevice()
+        public async Task TestInsertAlignedRecordsOfOneDevice()
         {
             var session_pool = new SessionPool(host, port, pool_size);
             await session_pool.Open(false);
@@ -239,30 +174,37 @@ namespace Apache.IoTDB.Samples
             System.Diagnostics.Debug.Assert(session_pool.IsOpen());
             var status = 0;
             await session_pool.DeleteStorageGroupAsync(test_group_name);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[1]),
-                TSDataType.BOOLEAN, TSEncoding.PLAIN, Compressor.SNAPPY);
+
+            string prefixPath = string.Format("{0}.{1}", test_group_name, test_device);
+            var measurement_lst = new List<string>()
+            {
+                test_measurements[1],
+                test_measurements[2],
+                test_measurements[3],
+                test_measurements[4],
+                test_measurements[5],
+                test_measurements[6]
+            };
+            var data_type_lst = new List<TSDataType>()
+            {
+                TSDataType.BOOLEAN, TSDataType.INT32, TSDataType.INT64, TSDataType.DOUBLE, TSDataType.FLOAT,
+                TSDataType.TEXT
+            };
+            var encoding_lst = new List<TSEncoding>()
+            {
+                TSEncoding.PLAIN, TSEncoding.PLAIN, TSEncoding.PLAIN, TSEncoding.PLAIN, TSEncoding.PLAIN,
+                TSEncoding.PLAIN
+            };
+            var compressor_lst = new List<Compressor>()
+            {
+                Compressor.SNAPPY, Compressor.SNAPPY, Compressor.SNAPPY, Compressor.SNAPPY, Compressor.SNAPPY,
+                Compressor.SNAPPY
+            };
+
+            status = await session_pool.CreateAlignedTimeseriesAsync(prefixPath, measurement_lst, data_type_lst, encoding_lst,
+                compressor_lst);
             System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[2]),
-                TSDataType.INT32, TSEncoding.PLAIN, Compressor.SNAPPY);
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[3]),
-                TSDataType.INT64, TSEncoding.PLAIN, Compressor.SNAPPY);
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[4]),
-                TSDataType.DOUBLE, TSEncoding.PLAIN, Compressor.SNAPPY);
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[5]),
-                TSDataType.FLOAT, TSEncoding.PLAIN, Compressor.SNAPPY);
-            System.Diagnostics.Debug.Assert(status == 0);
-            status = await session_pool.CreateTimeSeries(
-                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[6]), TSDataType.TEXT,
-                TSEncoding.PLAIN, Compressor.SNAPPY);
-            System.Diagnostics.Debug.Assert(status == 0);
+
             var device_id = string.Format("{0}.{1}", test_group_name, test_device);
             var measurements_lst = new List<List<string>>() { };
             measurements_lst.Add(new List<string>() { test_measurements[1], test_measurements[2] });
@@ -294,8 +236,7 @@ namespace Apache.IoTDB.Samples
                 var rowRecord = new RowRecord(timestamp_lst[i], values_lst[i], measurements_lst[i]);
                 rowRecords.Add(rowRecord);
             }
-
-            status = await session_pool.InsertRecordsOfOneDeviceAsync(device_id, rowRecords);
+            status = await session_pool.InsertAlignedRecordsOfOneDeviceAsync(device_id, rowRecords);
             System.Diagnostics.Debug.Assert(status == 0);
             var res = await session_pool.ExecuteQueryStatementAsync(
                 "select * from " + string.Format("{0}.{1}", test_group_name, test_device) + " where time<10");
@@ -303,7 +244,6 @@ namespace Apache.IoTDB.Samples
             while (res.HasNext()) Console.WriteLine(res.Next());
 
             await res.Close();
-            // large data test
             rowRecords = new List<RowRecord>() { };
             var tasks = new List<Task<int>>();
             for (var timestamp = 4; timestamp <= fetch_size * processed_size; timestamp++)
@@ -312,7 +252,7 @@ namespace Apache.IoTDB.Samples
                     new List<string>() { test_measurements[1], test_measurements[2] }));
                 if (timestamp % fetch_size == 0)
                 {
-                    tasks.Add(session_pool.InsertRecordsOfOneDeviceAsync(device_id, rowRecords));
+                    tasks.Add(session_pool.InsertAlignedRecordsOfOneDeviceAsync(device_id, rowRecords));
                     rowRecords = new List<RowRecord>() { };
                 }
             }
@@ -333,7 +273,7 @@ namespace Apache.IoTDB.Samples
             status = await session_pool.DeleteStorageGroupAsync(test_group_name);
             System.Diagnostics.Debug.Assert(status == 0);
             await session_pool.Close();
-            Console.WriteLine("TestInsertRecordsOfOneDevice Passed!");
+            Console.WriteLine("TestInsertAlignedRecordsOfOneDevice Passed!");
         }
     }
 }
