+++ samples/Apache.IoTDB.Samples/SessionPoolTest.Template.cs
@@ -21,7 +21,6 @@ namespace Apache.IoTDB.Samples
             MeasurementNode node2 = new MeasurementNode(test_measurements[2], TSDataType.INT64, TSEncoding.PLAIN, Compressor.SNAPPY);
             MeasurementNode node3 = new MeasurementNode(test_measurements[3], TSDataType.DOUBLE, TSEncoding.PLAIN, Compressor.SNAPPY);
             MeasurementNode node4 = new MeasurementNode(test_measurements[4], TSDataType.FLOAT, TSEncoding.PLAIN, Compressor.SNAPPY);
-            Console.WriteLine("node1 name : {0}", node1.Name);
 
             Template template = new Template(test_template_name);
             template.addToTemplate(node1);
@@ -106,13 +105,12 @@ namespace Apache.IoTDB.Samples
             System.Diagnostics.Debug.Assert(status == 0);
             status = await session_pool.AddAlignedMeasurementsInTemplateAsync(test_template_name, measurements);
             System.Diagnostics.Debug.Assert(status == 0);
-            var measurements_get = await session_pool.ShowMeasurementsInTemplateAsync(test_template_name);
-            foreach (var m in measurements_get)
+            var measurements_count = await session_pool.CountMeasurementsInTemplateAsync(test_template_name);
+            System.Diagnostics.Debug.Assert(measurements_count == 4);
+            foreach (var m in measurements)
             {
-                Console.WriteLine("measurement :\t{0}", m);
+                System.Diagnostics.Debug.Assert(await session_pool.IsMeasurementInTemplateAsync(test_template_name, m.Name));
             }
-            // status = await session_pool.SetSchemaTemplateAsync(test_template_name, string.Format("{0}.{1}", test_group_name, test_device));
-            // System.Diagnostics.Debug.Assert(status == 0);
             status = await session_pool.DropSchemaTemplateAsync(test_template_name);
             System.Diagnostics.Debug.Assert(status == 0);
             await session_pool.Close();
@@ -140,17 +138,47 @@ namespace Apache.IoTDB.Samples
             System.Diagnostics.Debug.Assert(status == 0);
             status = await session_pool.AddUnalignedMeasurementsInTemplateAsync(test_template_name, measurements);
             System.Diagnostics.Debug.Assert(status == 0);
-            var measurements_get = await session_pool.ShowMeasurementsInTemplateAsync(test_template_name);
-            foreach (var m in measurements_get)
+            var measurements_count = await session_pool.CountMeasurementsInTemplateAsync(test_template_name);
+            System.Diagnostics.Debug.Assert(measurements_count == 4);
+            foreach (var m in measurements)
             {
-                Console.WriteLine("measurement :\t{0}", m);
+                System.Diagnostics.Debug.Assert(await session_pool.IsPathExistInTemplate(test_template_name, m.Name));
             }
-            // status = await session_pool.SetSchemaTemplateAsync(test_template_name, string.Format("{0}.{1}", test_group_name, test_device));
-            // System.Diagnostics.Debug.Assert(status == 0);
             status = await session_pool.DropSchemaTemplateAsync(test_template_name);
             System.Diagnostics.Debug.Assert(status == 0);
             await session_pool.Close();
             Console.WriteLine("TestAddUnalignedMeasurements Passed!");
+        }
+        public async Task TestDeleteNodeInTemplate()
+        {
+            var session_pool = new SessionPool(host, port, pool_size);
+            await session_pool.Open(false);
+            if (debug) session_pool.OpenDebugMode();
+
+            System.Diagnostics.Debug.Assert(session_pool.IsOpen());
+            var status = 0;
+            await session_pool.DeleteStorageGroupAsync(test_group_name);
+            await session_pool.DropSchemaTemplateAsync(test_template_name);
+
+            MeasurementNode node1 = new MeasurementNode(test_measurements[1], TSDataType.INT32, TSEncoding.PLAIN, Compressor.SNAPPY);
+            MeasurementNode node2 = new MeasurementNode(test_measurements[2], TSDataType.INT64, TSEncoding.PLAIN, Compressor.SNAPPY);
+            MeasurementNode node3 = new MeasurementNode(test_measurements[3], TSDataType.DOUBLE, TSEncoding.PLAIN, Compressor.SNAPPY);
+            MeasurementNode node4 = new MeasurementNode(test_measurements[4], TSDataType.FLOAT, TSEncoding.PLAIN, Compressor.SNAPPY);
+            var measurements = new List<MeasurementNode>() { node1, node2, node3, node4 };
+
+            Template template = new Template(test_template_name);
+            status = await session_pool.CreateSchemaTemplateAsync(template);
+            System.Diagnostics.Debug.Assert(status == 0);
+            status = await session_pool.AddUnalignedMeasurementsInTemplateAsync(test_template_name, measurements);
+
+            status = await session_pool.DeleteNodeInTemplateAsync(test_template_name, test_measurements[1]);
+            var measurements_count = await session_pool.CountMeasurementsInTemplateAsync(test_template_name);
+            System.Diagnostics.Debug.Assert(measurements_count == 3);
+            System.Diagnostics.Debug.Assert(!await session_pool.IsPathExistInTemplate(test_template_name, test_measurements[1]));
+            status = await session_pool.DropSchemaTemplateAsync(test_template_name);
+            System.Diagnostics.Debug.Assert(status == 0);
+            await session_pool.Close();
+            Console.WriteLine("TestDeleteNodeInTemplate Passed!");
         }
     }
