+++ samples/Apache.IoTDB.Samples/SessionPoolTest.cs
@@ -1,0 +1,332 @@
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Apache.IoTDB.DataStructure;
+
+namespace Apache.IoTDB.Samples
+{
+    public partial class SessionPoolTest
+    {
+        public string host = "localhost";
+        public int port = 6667;
+        public string user = "root";
+        public string passwd = "root";
+        public int fetch_size = 500;
+        public int processed_size = 4;
+        public bool debug = false;
+        private int pool_size = 2;
+        public static string test_template_name = "TEST_CSHARP_CLIENT_TEMPLATE_97209";
+        public static string test_group_name = "root.TEST_CSHARP_CLIENT_GROUP_97209";
+        public static string test_device = "TEST_CSHARP_CLIENT_DEVICE";
+        public static string test_measurement = "TEST_CSHARP_CLIENT_TS";
+        public static List<int> device_count = new List<int>() { 0, 1, 2, 3 };
+        public static List<int> measurement_count = new List<int>() { 0, 1, 2, 3, 4, 5, 6 };
+        public static List<string> test_devices = new List<string>(
+            device_count.ConvertAll(x => test_device + x.ToString()).ToArray()
+        );
+        public List<string> test_measurements = new List<string>(
+            measurement_count.ConvertAll(x => test_measurement + x.ToString()).ToArray()
+        );
+
+
+        public SessionPoolTest(string _host = "localhost")
+        {
+            host = _host;
+        }
+
+        public void Test()
+        {
+            Task task;
+
+            task = TestInsertAlignedRecord();
+            task.Wait();
+            task = TestInsertAlignedRecords();
+            task.Wait();
+            task = TestInsertAlignedRecordsOfOneDevice();
+            task.Wait();
+            task = TestInsertAlignedTablet();
+            task.Wait();
+            task = TestInsertAlignedTablets();
+            task.Wait();
+            task = TestInsertRecord();
+            task.Wait();
+            task = TestCreateMultiTimeSeries();
+            task.Wait();
+            task = TestInsertStrRecord();
+            task.Wait();
+            task = TestInsertRecords();
+            task.Wait();
+            task = TestInsertRecordsOfOneDevice();
+            task.Wait();
+            task = TestInsertTablet();
+            task.Wait();
+            task = TestInsertTablets();
+            task.Wait();
+            task = TestAddAlignedMeasurements();
+            task.Wait();
+            task = TestAddUnalignedMeasurements();
+            task.Wait();
+            task = TestSetAndUnsetSchemaTemplate();
+            task.Wait();
+            task = TestCreateAlignedTimeseries();
+            task.Wait();
+            task = TestCreateAndDropSchemaTemplate();
+            task.Wait();
+            task = TestDeleteNodeInTemplate();
+            task.Wait();
+            task = TestGetTimeZone();
+            task.Wait();
+            task = TestSetAndDeleteStorageGroup();
+            task.Wait();
+            task = TestCreateTimeSeries();
+            task.Wait();
+            task = TestDeleteTimeSeries();
+            task.Wait();
+            task = TestDeleteStorageGroups();
+            task.Wait();
+            task = TestCheckTimeSeriesExists();
+            task.Wait();
+            task = TestSetTimeZone();
+            task.Wait();
+            task = TestDeleteData();
+            task.Wait();
+            task = TestNonSql();
+            task.Wait();
+            task = TestSqlQuery();
+            task.Wait();
+        }
+
+        public async Task TestGetTimeZone()
+        {
+            var session_pool = new SessionPool(host, port, pool_size);
+            await session_pool.Open(false);
+            if (debug) session_pool.OpenDebugMode();
+
+            await session_pool.DeleteStorageGroupAsync(test_group_name);
+            System.Diagnostics.Debug.Assert(session_pool.IsOpen());
+            var time_zone = await session_pool.GetTimeZone();
+            System.Diagnostics.Debug.Assert(time_zone == "UTC+08:00");
+            await session_pool.Close();
+            Console.WriteLine("TestGetTimeZone Passed!");
+        }
+
+
+
+        public async Task TestSetAndDeleteStorageGroup()
+        {
+            var session_pool = new SessionPool(host, port, pool_size);
+            var status = 0;
+            await session_pool.Open(false);
+            if (debug) session_pool.OpenDebugMode();
+
+            status = await session_pool.DeleteStorageGroupAsync(test_group_name);
+            System.Diagnostics.Debug.Assert(
+                await session_pool.SetStorageGroup(test_group_name) == 0);
+            System.Diagnostics.Debug.Assert(
+                await session_pool.DeleteStorageGroupAsync(test_group_name) == 0);
+            await session_pool.Close();
+            Console.WriteLine("TestSetAndDeleteStorageGroup Passed!");
+        }
+
+
+        public async Task TestDeleteStorageGroups()
+        {
+            var session_pool = new SessionPool(host, port, pool_size);
+            await session_pool.Open(false);
+            if (debug) session_pool.OpenDebugMode();
+
+            await session_pool.SetStorageGroup(string.Format("{0}{1}", test_group_name, "_01"));
+            await session_pool.SetStorageGroup(string.Format("{0}{1}", test_group_name, "_02"));
+            await session_pool.SetStorageGroup(string.Format("{0}{1}", test_group_name, "_03"));
+            await session_pool.SetStorageGroup(string.Format("{0}{1}", test_group_name, "_04"));
+            var group_names = new List<string>() { };
+            group_names.Add(string.Format("{0}{1}", test_group_name, "_01"));
+            group_names.Add(string.Format("{0}{1}", test_group_name, "_02"));
+            group_names.Add(string.Format("{0}{1}", test_group_name, "_03"));
+            group_names.Add(string.Format("{0}{1}", test_group_name, "_04"));
+            System.Diagnostics.Debug.Assert(await session_pool.DeleteStorageGroupsAsync(group_names) == 0);
+            await session_pool.Close();
+            Console.WriteLine("TestDeleteStorageGroups Passed!");
+        }
+
+
+        public async Task TestSetTimeZone()
+        {
+            var session_pool = new SessionPool(host, port, pool_size);
+            await session_pool.Open(false);
+            if (debug) session_pool.OpenDebugMode();
+
+            await session_pool.DeleteStorageGroupAsync(test_group_name);
+            System.Diagnostics.Debug.Assert(session_pool.IsOpen());
+            await session_pool.SetTimeZone("GMT+8:00");
+            System.Diagnostics.Debug.Assert(await session_pool.GetTimeZone() == "GMT+8:00");
+            await session_pool.Close();
+            Console.WriteLine("TestSetTimeZone Passed!");
+        }
+
+        public async Task TestDeleteData()
+        {
+            var session_pool = new SessionPool(host, port, pool_size);
+            var status = 0;
+            await session_pool.Open(false);
+            if (debug) session_pool.OpenDebugMode();
+
+            System.Diagnostics.Debug.Assert(session_pool.IsOpen());
+            status = await session_pool.DeleteStorageGroupAsync(test_group_name);
+
+            status = await session_pool.CreateTimeSeries(
+                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[1]), TSDataType.TEXT,
+                TSEncoding.PLAIN, Compressor.UNCOMPRESSED);
+            System.Diagnostics.Debug.Assert(status == 0);
+            status = await session_pool.CreateTimeSeries(
+                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[2]),
+                TSDataType.BOOLEAN, TSEncoding.PLAIN, Compressor.UNCOMPRESSED);
+            System.Diagnostics.Debug.Assert(status == 0);
+            status = await session_pool.CreateTimeSeries(
+                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[3]),
+                TSDataType.INT32, TSEncoding.PLAIN, Compressor.UNCOMPRESSED);
+            System.Diagnostics.Debug.Assert(status == 0);
+
+            var measures = new List<string>
+            {
+                test_measurements[1], test_measurements[2], test_measurements[3]
+            };
+            var values = new List<object> { "test_text", true, (int)123 };
+            status = await session_pool.InsertRecordAsync(
+                string.Format("{0}.{1}", test_group_name, test_device), new RowRecord(1, values, measures));
+            System.Diagnostics.Debug.Assert(status == 0);
+            status = await session_pool.InsertRecordAsync(
+                string.Format("{0}.{1}", test_group_name, test_device), new RowRecord(2, values, measures));
+            System.Diagnostics.Debug.Assert(status == 0);
+            status = await session_pool.InsertRecordAsync(
+                string.Format("{0}.{1}", test_group_name, test_device), new RowRecord(3, values, measures));
+            System.Diagnostics.Debug.Assert(status == 0);
+            status = await session_pool.InsertRecordAsync(
+                string.Format("{0}.{1}", test_group_name, test_device), new RowRecord(4, values, measures));
+            System.Diagnostics.Debug.Assert(status == 0);
+            var res = await session_pool.ExecuteQueryStatementAsync(
+                "select * from " + string.Format("{0}.{1}", test_group_name, test_device) + " where time<10");
+            res.ShowTableNames();
+            while (res.HasNext()) Console.WriteLine(res.Next());
+
+            await res.Close();
+            var ts_path_lst = new List<string>()
+            {
+                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[1]),
+                string.Format("{0}.{1}.{2}", test_group_name, test_device, test_measurements[2]),
+            };
+            await session_pool.DeleteDataAsync(ts_path_lst, 2, 3);
+            res = await session_pool.ExecuteQueryStatementAsync(
+                "select * from " + string.Format("{0}.{1}", test_group_name, test_device) + " where time<10");
+            res.ShowTableNames();
+            while (res.HasNext()) Console.WriteLine(res.Next());
+
+            await res.Close();
+            status = await session_pool.DeleteStorageGroupAsync(test_group_name);
+            System.Diagnostics.Debug.Assert(status == 0);
+            await session_pool.Close();
+            Console.WriteLine("TestDeleteData Passed!");
+        }
+
+        public async Task TestNonSql()
+        {
+            var session_pool = new SessionPool(host, port, pool_size);
+            var status = 0;
+            await session_pool.Open(false);
+            if (debug) session_pool.OpenDebugMode();
+
+            System.Diagnostics.Debug.Assert(session_pool.IsOpen());
+            status = await session_pool.DeleteStorageGroupAsync(test_group_name);
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "create timeseries " + string.Format("{0}.{1}", test_group_name, test_device) + ".status with datatype=BOOLEAN,encoding=PLAIN");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "create timeseries " + string.Format("{0}.{1}", test_group_name, test_device) + ".temperature with datatype=FLOAT,encoding=PLAIN");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "create timeseries " + string.Format("{0}.{1}", test_group_name, test_device) + ".hardware with datatype=TEXT,encoding=PLAIN");
+            status = await session_pool.ExecuteNonQueryStatementAsync(
+                "insert into " + string.Format("{0}.{1}", test_group_name, test_device) + "(timestamp, status, temperature, hardware) VALUES (4, false, 20, 'yxl')");
+            System.Diagnostics.Debug.Assert(status == 0);
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "insert into " + string.Format("{0}.{1}", test_group_name, test_device) + "(timestamp, status, temperature, hardware) VALUES (5, true, 12, 'myy')");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "insert into " + string.Format("{0}.{1}", test_group_name, test_device) + "(timestamp, status, temperature, hardware) VALUES (6, true, 21, 'lz')");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "insert into " + string.Format("{0}.{1}", test_group_name, test_device) + "(timestamp, status, hardware) VALUES (7, true,'lz')");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "insert into " + string.Format("{0}.{1}", test_group_name, test_device) + "(timestamp, status, hardware) VALUES (7, true,'lz')");
+            var res = await session_pool.ExecuteQueryStatementAsync(
+                "select * from " + string.Format("{0}.{1}", test_group_name, test_device) + " where time<10");
+            res.ShowTableNames();
+            while (res.HasNext()) Console.WriteLine(res.Next());
+
+            await res.Close();
+            status = await session_pool.DeleteStorageGroupAsync(test_group_name);
+            System.Diagnostics.Debug.Assert(status == 0);
+            await session_pool.Close();
+            Console.WriteLine("TestNonSql Passed");
+        }
+
+        public async Task TestSqlQuery()
+        {
+            var session_pool = new SessionPool(host, port, pool_size);
+            var status = 0;
+            await session_pool.Open(false);
+            if (debug) session_pool.OpenDebugMode();
+
+            System.Diagnostics.Debug.Assert(session_pool.IsOpen());
+            status = await session_pool.DeleteStorageGroupAsync(test_group_name);
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "create timeseries " + string.Format("{0}.{1}", test_group_name, test_device) + ".status with datatype=BOOLEAN,encoding=PLAIN");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "create timeseries " + string.Format("{0}.{1}", test_group_name, test_device) + ".temperature with datatype=FLOAT,encoding=PLAIN");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "create timeseries " + string.Format("{0}.{1}", test_group_name, test_device) + ".hardware with datatype=TEXT,encoding=PLAIN");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "insert into " + string.Format("{0}.{1}", test_group_name, test_device) + "(timestamp, status, temperature, hardware) VALUES (4, false, 20, 'yxl')");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "insert into " + string.Format("{0}.{1}", test_group_name, test_device) + "(timestamp, status, temperature, hardware) VALUES (5, true, 12, 'myy')");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "insert into " + string.Format("{0}.{1}", test_group_name, test_device) + "(timestamp, status, temperature, hardware) VALUES (6, true, 21, 'lz')");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "insert into " + string.Format("{0}.{1}", test_group_name, test_device) + "(timestamp, status, hardware) VALUES (7, true,'lz')");
+            await session_pool.ExecuteNonQueryStatementAsync(
+                "insert into " + string.Format("{0}.{1}", test_group_name, test_device) + "(timestamp, status, hardware) VALUES (7, true,'lz')");
+
+            var res = await session_pool.ExecuteQueryStatementAsync("show timeseries root");
+            res.ShowTableNames();
+            while (res.HasNext()) Console.WriteLine(res.Next());
+
+            await res.Close();
+            Console.WriteLine("SHOW TIMESERIES ROOT sql passed!");
+            res = await session_pool.ExecuteQueryStatementAsync("show devices");
+            res.ShowTableNames();
+            while (res.HasNext()) Console.WriteLine(res.Next());
+
+            await res.Close();
+            Console.WriteLine("SHOW DEVICES sql passed!");
+            res = await session_pool.ExecuteQueryStatementAsync("COUNT TIMESERIES root");
+            res.ShowTableNames();
+            while (res.HasNext()) Console.WriteLine(res.Next());
+
+            await res.Close();
+            Console.WriteLine("COUNT TIMESERIES root sql Passed");
+            res = await session_pool.ExecuteQueryStatementAsync("select * from root.ln.wf01 where time<10");
+            res.ShowTableNames();
+            while (res.HasNext()) Console.WriteLine(res.Next());
+
+            await res.Close();
+            Console.WriteLine("SELECT sql Passed");
+            res = await session_pool.ExecuteQueryStatementAsync(
+                "select * from " + string.Format("{0}.{1}", test_group_name, test_device) + " where time<10");
+            res.ShowTableNames();
+            while (res.HasNext()) Console.WriteLine(res.Next());
+
+            await res.Close();
+            status = await session_pool.DeleteStorageGroupAsync(test_group_name);
+            System.Diagnostics.Debug.Assert(status == 0);
+            await session_pool.Close();
+            Console.WriteLine("SELECT sql Passed");
+        }
+    }
+}
